tests/wpt/web-platform-tests/css/css-text-decor/text-decoration-width-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Decoration Test: parsing text-decoration-width with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-4/#text-decoration-width-property">
<meta name="assert" content="text-decoration-width supports the following values: auto | from-font | <length>">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("text-decoration-width", "otto");
test_invalid_value("text-decoration-width", "asdlflj");
test_invalid_value("text-decoration-width", "-10");
test_invalid_value("text-decoration-width", "60002020");
test_invalid_value("text-decoration-width", "!@#$%^&");
test_invalid_value("text-decoration-width", "10e2");
test_invalid_value("text-decoration-width", "from font");
test_invalid_value("text-decoration-width", "10%");
test_invalid_value("text-decoration-width", "-27%");
test_invalid_value("text-decoration-width", "calc(40% - 20px)");
test_invalid_value("text-decoration-width", "calc(100% - 40em)");
</script>
</body>
</html>
